+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.3.1 under development
 
-- no changes in this release.
+- Bug #224: Fix typo in `UNAVAILABLE` exit code reason (@vjik)
 
 ## 2.3.0 January 23, 2025