+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.3.1 under development
 
-- no changes in this release.
+- Enh #226: Check empty string as key in command map in `CommandLoader` validation (@vjik)
 
 ## 2.3.0 January 23, 2025