+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Enh #226: Check empty string as key in command map in `CommandLoader` validation (@vjik)
 - Bug #224: Fix typo in `UNAVAILABLE` exit code reason (@vjik)
 - Bug #227: Throw `RuntimeException` in `ServeCommand` when it wasn't possible to find out the current directory (@vjik)
+- Bug #230: The `serve` command no longer opens the browser by default (@vjik)
 
 ## 2.3.0 January 23, 2025